backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
docker==6.1.3
gitpython==3.1.40
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
cryptography==41.0.7 
email-validator==2.1.0
celery==5.3.4
redis==5.0.1
kombu==5.3.4 